+++ src/domain/smalltalk/service/SmallTalk.Service.ts
@@ -91,4 +91,5 @@ export class SmallTalkService{
         }
         return true;
     }
+
 }
